Stop the expired period-item tooltip from crashing the client
ExpireTime was filled in only on the not-yet-expired branch
(bExpiredPeriod == false) but read further down whenever lExpireTime != 0 -
including for an already-expired item, which is the normal end state of every
period item. So drawing that tooltip dereferenced whatever happened to be on
the stack. It is now computed whenever there is a timestamp to convert,
matching where it is actually used, and localtime() returning NULL for a bad
timestamp is handled rather than trusted.

PMSG_PERIODITEMEX_ITEMLIST's lExpireDate was also a bare `long`: 4 bytes on the
PC build, 8 under Android's LP64 ABI, so every field the client read past it on
the wire was garbage there. `int` is 4 bytes on both, matching the wire format
the PC build already produced - no change on PC, corrective on Android. Same
trap as the cash-shop packet structs.

Confirmed working on device.

+++ Source_Server-Mobi/5.Main/source/WSclient.h
@@ -3702,7 +3702,10 @@ typedef struct
 	
 	WORD				wItemCode;
 	WORD				wItemSlotIndex;
-	long				lExpireDate;
+	int					lExpireDate;	// was `long`: 4 bytes on the server/PC, 8 on 64-bit
+										// Android, so the field read off the wire past this
+										// point was garbage there. `int` is 4 bytes on every
+										// target this ships to.
 }PMSG_PERIODITEMEX_ITEMLIST, *LPPMSG_PERIODITEMEX_ITEMLIST;
 
 #endif // KJH_ADD_PERIOD_ITEM_SYSTEM
